Add unstable flag to extend compiletest
`auto_enum` uses the hash value of the input AST to prevent access to
the generated enum.
This works well for common use cases, but is inconvenient when testing
error messages that contain enum names.
This PR introduces an unstable cfg flag that depends on the unstable
`proc_macro_def_site` feature. When this feature is enabled, `auto_enum`
uses the enum name is based on the function name, and access to the enum
is prevented by using `Span::def_site()`.
Currently, this feature only affects when `auto_enum` is used in
function position.