Forbid custom Unpin/Drop impl if trait has `Pin<&mut self>` receiver
If a user creates their own Unpin or Drop implementation, trait
implementations with `Pin<&mut self>` receiver can cause unsoundness.

This was not a problem in #[auto_enum] attribute where enums are
anonymized, but it becomes a problem when users have access to enums
(i.e., when using #[enum_derive]).

So, we ensure safety here by an Unpin implementation that implements
Unpin only if all fields are Unpin (this also forbids custom Unpin
implementation), and a hack that forbids custom Drop implementation.
(Both are what pin-project does by default.) The repr(packed) check is
not needed since repr(packed) is not available on enum.